+++ structs.lua
@@ -1,8 +1,11 @@
 return {
   DataMatrix = require("balm/s/data_matrix"),
+  Deque = require("balm/s/deque"),
   LinkedList = require("balm/s/linked_list"),
   List = require("balm/s/list"),
   MinHeap = require("balm/s/min_heap"),
   Records = require("balm/s/records"),
   RingBuffer = require("balm/s/ring_buffer"),
+  Vector = require("balm/s/vector"),
+  WeightedList = require("balm/s/weighted_list"),
 }